Unify FeedViewController static factory helper into the `FeedUIComposer` and move to composer
